travel time

+++ src/travel_time_computation/travel_time_Baltimore.py
@@ -1,40 +1,59 @@
 import os
-import datetime
-from src.util import config
+from datetime import datetime, timedelta
+from src.util import get_config
 from travel_time_computation import compute_travel_time_matrices
 
-config = config.get_config()
+config = get_config()
 
 raw_data_path = config['raw_data_path']
 block_group_shapefile_path = raw_data_path + config['block_group_shapefile_path']
 md_rac_path = raw_data_path + config['rac_file_path']
-md_wac_path = raw_data_path +  + config['wac_file_path']
+md_wac_path = raw_data_path + config['wac_file_path']
 osm_path = raw_data_path + config['osm_path']
 # Filepath to GTFS data
-date = "2023-03-06"
-departure_time = datetime.datetime(2023, 3, 6, 8, 30)
-GTFS = [f"processed_data/updated_gtfs/{date}.zip", 
-        'raw_data/rail_gtfs.zip', 
-        'raw_data/subway_gtfs.zip',
-        'raw_data/commuterbus_gtfs.zip', 
-        'raw_data/train_gtfs.zip'
-        ]
-# Filepath to newline GTFS data
-redline = "processed_data/redline.zip"
 
-wait_minutes = 30
+begin_date = config['begin_date']
+end_date = config['end_date']
 
-from prepare_data import prepare_data_Baltimore
+# Define the step (in this case, one day)
+step = timedelta(days=1)
 
-md_rac_df, md_wac_df = prepare_data_Baltimore(block_group_shapefile_path, md_rac_path, md_wac_path)
+# Iterate through dates from begin_date to end_date
+current_date = begin_date
 
-matrix_before_redline, matrix_after_redline = \
-    compute_travel_time_matrices(md_rac_df, md_wac_df, departure_time, osm_path, GTFS, redline, wait_minutes=wait_minutes)
+while current_date <= end_date:
 
-os.makedirs('processed_data/travel_time_matrices', exist_ok=True)
-matrix_before_redline.to_csv(f"processed_data/travel_time_matrices/{date}_before_redline.csv", index = False)
-matrix_after_redline.to_csv(f"processed_data/travel_time_matrices/{date}_after_redline.csv", index = False)
+    month = int(current_date.strftime('%m'))
+    day = int(current_date.strftime('%d'))
+    date = current_date.strftime('%Y-%m-%d')
+    current_date += step
 
-travel_time = matrix_before_redline.merge(matrix_after_redline, on=['from_id', 'to_id'], suffixes=('_before', '_after'))
+    bus_gtfs_path = f'processed_data/updated_gtfs/{date}.zip'
+    if os.path.exists(bus_gtfs_path):
+        print(date)
+        departure_time = datetime(2023, month, day, 8, 0)
+        GTFS = [f"processed_data/updated_gtfs/{date}.zip", 
+                'raw_data/rail_gtfs.zip', 
+                'raw_data/subway_gtfs.zip',
+                'raw_data/commuterbus_gtfs.zip', 
+                'raw_data/train_gtfs.zip'
+                ]
+        # Filepath to newline GTFS data
+        redline = "processed_data/redline.zip"
 
-travel_time.to_csv(f"processed_data/travel_time_matrices/{date}_travel_time.csv", index = False)
+        wait_minutes = 30
+
+        from prepare_data import prepare_data_Baltimore
+
+        md_rac_df, md_wac_df = prepare_data_Baltimore(block_group_shapefile_path, md_rac_path, md_wac_path)
+
+        matrix_before_redline, matrix_after_redline = \
+            compute_travel_time_matrices(md_rac_df, md_wac_df, departure_time, osm_path, GTFS, redline, wait_minutes=wait_minutes)
+
+        os.makedirs('processed_data/travel_time_matrices', exist_ok=True)
+        matrix_before_redline.to_csv(f"processed_data/travel_time_matrices/{date}_before_redline.csv", index = False)
+        matrix_after_redline.to_csv(f"processed_data/travel_time_matrices/{date}_after_redline.csv", index = False)
+
+        travel_time = matrix_before_redline.merge(matrix_after_redline, on=['from_id', 'to_id'], suffixes=('_before', '_after'))
+
+        travel_time.to_csv(f"processed_data/travel_time_matrices/{date}_travel_time.csv", index = False)
